fix:default title of the chart should be updated to the graphed value… (#8602)
fixes: #8579
/claim #8579 

## Summary
case 1: If the graph is created/edited without any title it will pick
from function type and metric. [functionType(metric)]
case 2: If the user wants to give a proper name to it, they can fill in
the title of their choice. The user's title will be given a higher
priority irrespective of the updation of function and metric.

note: If the user wants the system to pick the title. Make the title
empty. It will automatically pick the title based on function type and
metric.
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->
https://www.loom.com/share/9fd9a1e000d84c5098afe7330f0dc6e5

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->
No

## Does this work require review from our design team?
No

<!--
 Request review from julian-highlight / our design team 
-->