update sendResponse snippet in cloudflare sdk documentation (#8746)
## Summary

There seemed to be 2 issues with the snippet for `sendResponse` on the
Cloudflare SDK documentation.

1. `return` when creating the response would prevent any following code
from running
2. The function signature for `sendResponse` only takes in one argument,
the response object. Not all 4 arguments being passed in the snippet.

From index.d.ts:
```ts
interface HighlightInterface {
    init: (request: Request, env: HighlightEnv, ctx: ExecutionContext, service?: string) => WorkersSDK;
    isInitialized: () => boolean;
    consumeError: (error: Error) => void;
    sendResponse: (response: Response) => void;
    setAttributes: (attributes: ResourceAttributes) => void;
    recordMetric: (metric: Metric) => void;
}
```

## How did you test this change?

This is a documentation change, so no testing was done. Please verify as
you can.

## Are there any deployment considerations?

No migration or deployment concerns.

## Does this work require review from our design team?

I do not think so.