HIG-4768 add new session exclusion reasons to ui (#8886)
## Summary
- some of the newer exclusion reasons would show the default "session
does not exist or has not been made public" error, which is confusing
when a user gets linked to an excluded session e.g. from an alert
- update the logic to have all exclusion types
- relax `payload_updated_at` not-null requirement as some of these
sessions do not have this set (e.g. if no events)
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?

<!--
 Request review from julian-highlight / our design team 
-->