Fix network request trace view (#9122)
## Summary

When we enabled browser OTEL data collection we started collecting
traces for network requests. These were being recorded alongside the
traces we were already creating with out XHR and fetch listeners. Traces
on the backend are correctly being associated with the browser OTEL
traces, but the network resources in a session are still using the
request ID generated on the client as their trace ID, creating a new
trace that is not connected to the OTEL trace or any of the server code.

This PR sets the request ID to the active span's trace ID, if available.
It also prevents network request spans from being created if we detect a
`traceparent`, which would mean that we already have an OTEL-based trace
from the browser instrumentation.

## How did you test this change?

* Start a new session and click around to ensure the session has some activity
* View the session in the player
* View the network request panel
* Open a network request and ensure the backend spans are connected to the trace

## Are there any deployment considerations?

Will click test to confirm things are fixed on new sessions in our project, which should be the only project with browser otel enabled so far.

## Does this work require review from our design team?

N/A