fix a couple crashes from undefined series (#9678)
## Summary
- fixes a few errors:
`Cannot read properties of undefined (reading 'column')`
`Cannot read properties of undefined (reading 'series')`
- remove some debug code
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- repro'd crashing locally, confirmed fixed
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->