fix [object Object] in graph csv export (#9756)
## Summary
- splits the session csv and graph csv `processRows` logic into 2 to
account for the new format for graph data
- `useGraphContext` to make sure context is shared between all
`DashboardCell` instances now
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally with different graph permutations
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->