fix alert graph loop, graphs with time range and no rows (#9912)
## Summary
- non-memoized `expressions` was causing graph to be redrawn repeatedly
and lagging / crashing the UI
- no data (e.g. an alert that did not trigger in the shown time period)
was displaying as unix zero time
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- should refactor useEffect in `Graph` to prevent non-memoized arguments
from doing this in the future
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->