SQL builder - support more complex queries (#9875)
## Summary
- instead of transforming the input into a sqlbuilder select query, keep
it in parsed form
- this allows more flexible types of input query (e.g. CTEs, union
queries, nested selects)
- identify parts querying the main product tables and apply default
filters to those only
- add other filters to the parsed form before transforming back into a
string
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally for multiple resource types, complex queries
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->