fix events, add timestamp override (#9928)
## Summary
- fix issue w/ event table references
- if timestamp is used in the where clause, override the default
timestamp (this is helpful for calculating backward-looking things like
MAU)
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- https://www.loom.com/share/0d6338c12de144cd8837d0eb1e7f9d23
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->