support join for non-resource tables, fix issues w/ OVER ROWS expressions (#9975)
## Summary
- support joins for non-resource tables, fix certain window expressions
by forking the parser
- eventually we can support joins for resource tables, but the default
filtering logic is difficult to apply
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally w/ window function
![Screenshot 2025-02-12 at 11 26
20 AM](https://github.com/user-attachments/assets/2d15d557-f860-465e-8361-068abaab2ac3)

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->