Check error retention on last error object (#10016)
## Summary
Error groups are erroring out because the retention policy is checking
the updated at date. However, this does not reflect when the latest
error object was created. This value can also be reset by setting the
status of the error group allowing the error group to have a retention
past its last error instance.

Update the check to use the last error instance time.

## How did you test this change?
1. Load an error group
2. Show be able to view the error instance

## Are there any deployment considerations?
Check for reported customers

## Does this work require review from our design team?
N/A