remove extra marshalling/unmarshalling in pushPayload (#10049)
## Summary
- there are a few places in pushPayload that we marshal/unmarshal to
transform between types; however this is expensive and circumvents the
type system
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- clicktested locally to make sure sessions were being recorded
correctly
- observed logs to check for 
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- will monitor prod after deployment - able to revert and replay the
sessions queue if issues
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->