increase ingest sampling rate to 50% (#10077)
## Summary
- increase ingest sampling rate to 50%
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- will monitor production traffic
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- 10% sampling is still in effect for pushPayload processing. Increase
ingest rate first, then increase processing sampling if all looks good
- no, can revert if issues
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->