push payload perf (#10089)
## Summary
- in production, the
`public-graph.pushPayload.parseEvents.remarshalEvents` span is 30%-40%
of the total processPayload time for larger messages
- this change switches every `resolver.go` json.Marshal / Unmarshal call
to use an alternate library
- change sampling logic to sample in 1/1000 of processPayload messages
to ingest more processPayload spans and avoid overrepresenting larger
messages
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- ran locally, saw an improvement in the remarshal spans
![Screenshot 2025-03-17 at 2 08
48 PM](https://github.com/user-attachments/assets/0b06c145-8b36-4ed7-b380-5aec5760811c)

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- no, can revert if issues 
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

## Does this work require review from our design team?
- no
<!--
 Request review from julian-highlight / our design team 
-->