turn off session cookie storage by default (#9984)
## Summary

Replace the `skipCookieSessionDataLoad` setting with an opt-in version
(`sessionCookie: true`)
to default to not using the session cookie. While the cookie technique
is useful to associate
server-side rendering context with the frontend session replay which
starts after the trace,
the cookies cause HTTP request bloat when multiple navigations occur
since they are not cleared.

Recommend the setting be used only when needed (for .NET Blazer apps,
for instance).

## How did you test this change?

CI

## Are there any deployment considerations?

changeset included

## Does this work require review from our design team?

no